backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
supabase>=2.11.0
email-validator==2.1.0.post1
gunicorn==21.2.0
uvicorn[standard]==0.24.0
